[TASK] Add MediaFactory skeleton
I chose the term "Media" instead of "File" to match the usage
in Flow\Utility\MediaTypes.

Hopefully the term "Media" suggests that you should make a MediaFactory
for each kind of file you need to handle in the builder. In the future, a
builder might choose not to build things that are stored as files. I can imagine
persisting some builder derivatives in a document database or some other mechanism.