[WIP][TASK] Add Irregular Package Manager
This should help builders deal with derivatives that are not flow
packages.

An Irregular Package is a package that does not follow Flow
conventions, or which needs to be treated in a special way beyond the
control of the Flow Package Manager.

Even though some TYPO3CMS extensions follow some Flow organization
principles, especially those that use extbase/fluid,
they are not Flow packages, and would be handled as Irregular
packages.

Typically Irregular Packages are not stored under Packages/. Instead
a subdirectory of Data/ is preferred. However, this also provides the
option of working with packages in Packages/Libraries/.

A builder package that wishes to work with Irregular Packages will
need to create an IrregularPackageStrategy and configure the
IrregularPackageManager to use that strategy. The Package Builder,
for example, will need to add a TYPO3CMS Extension strategy. That
strategy might be provided by another package so that it can be
reused by multiple builders.

TODO: This is still very broken.